MES 23/09/2026

+++ src/helpers/attendanceSheets.js
@@ -82,7 +82,14 @@ exports.create = async (payload, credentials) => {
         const signatureCopy = cloneDeep(payload.signature);
         if (get(formationExpoTokenList, 'length')) formationExpoTokens[trainee] = formationExpoTokenList;
         const attendanceSheet = await AttendanceSheet
-          .findOne({ trainee, course: payload.course, slots: { $exists: true }, file: { $exists: false } }).lean();
+          .findOne({
+            trainee,
+            course: payload.course,
+            trainer: payload.trainer,
+            slots: { $exists: true },
+            file: { $exists: false },
+          })
+          .lean();
         let slotWithTrainerSignature = null;
         if (attendanceSheet && course.type === INTER_B2B) {
           slotWithTrainerSignature = attendanceSheet.slots
@@ -203,7 +210,9 @@ exports.update = async (attendanceSheetId, payload, credentials) => {
       promises.push(AttendanceHelper.create({ trainee, courseSlot: slot }, credentials));
     }
   }
-  promises.push(Attendance.deleteMany({ courseSlot: { $in: attendancesToDelete }, trainee }));
+  if (payload.shouldDeleteAttendances) {
+    promises.push(Attendance.deleteMany({ courseSlot: { $in: attendancesToDelete }, trainee }));
+  }
   promises.push(
     AttendanceSheet.updateOne({ _id: attendanceSheetId }, { $set: { slots: payload.slots.map(s => ({ slotId: s })) } })
   );
